feat: separate LanguageContext from LanguageProvider components
