+++ host-app/src/context/LanguageContext.ts
@@ -1,0 +1,10 @@
+import { createContext } from 'react';
+
+export type Language = 'en' | 'es';
+
+export interface LanguageContextProps {
+  language: Language;
+  setLanguage: (lang: Language) => void;
+}
+
+export const LanguageContext = createContext<LanguageContextProps | undefined>(undefined);
