--- host-app/src/context/LanguageContext.tsx
@@ -1,21 +1,0 @@
-import { createContext, ReactNode, useState } from 'react';
-
-export type Language = 'en' | 'es';
-
-interface LanguageContextProps {
-  language: Language;
-  setLanguage: (lang: Language) => void;
-}
-
-export const LanguageContext = createContext<LanguageContextProps | undefined>(undefined);
-
-export const LanguageProvider = ({ children }: { children: ReactNode }) => {
-  const [language, setLanguage] = useState<Language>('en');
-  return (
-    <LanguageContext.Provider value={{ language, setLanguage }}>
-      {children}
-    </LanguageContext.Provider>
-  );
-};
-
-
