+++ host-app/src/context/LanguageProvider.tsx
@@ -1,0 +1,11 @@
+import { ReactNode, useState } from 'react';
+import { Language, LanguageContext } from './LanguageContext';
+
+export const LanguageProvider = ({ children }: { children: ReactNode }) => {
+  const [language, setLanguage] = useState<Language>('en');
+  return (
+    <LanguageContext.Provider value={{ language, setLanguage }}>
+      {children}
+    </LanguageContext.Provider>
+  );
+};
